[MoveOnly] Move non-wallet RPC files to subdir
