+++ src/rpc/blockchain.cpp
@@ -8,7 +8,7 @@
 
 #include "checkpoints.h"
 #include "main.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "sync.h"
 #include "util.h"
 #include "utilmoneystr.h"
