+++ src/rpc/misc.cpp
@@ -13,7 +13,7 @@
 #include "masternode-sync.h"
 #include "net.h"
 #include "netbase.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "timedata.h"
 #include "util.h"
