+++ src/rpc/protocol.cpp
@@ -6,7 +6,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "rpcprotocol.h"
+#include "rpc/protocol.h"
 
 #include "random.h"
 #include "tinyformat.h"
